requirements.txt
streamlit_authenticator==0.2.2
streamlit==1.24.1
openpyxl==3.0.10
plotly==4.14.3
deta==1.1.0
pandas==1.4.3
python-dotenv==0.21.0